+++ src/commands/CommandCommonInfo.ts
@@ -4,3 +4,4 @@ export const DOCLINK_COMMAND = 'jsonHelper.docLink';
 export const COPYTOCLIPBOARD_COMMAND = 'jsonHelper.copyToClipboard';
 export const MOVE_PRE_CMD = 'jsonHelper.moveToPreviousKey';
 export const MOVE_NEXT_CMD = 'jsonHelper.moveToNextKey';
+export const SHOW_NODES_QUICK_PICK_CMD = 'jsonHelper.showNodesInQuickPick';